requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
goose3==3.1.19
nltk==3.8.1
openai==1.40.1
pdfplumber==0.11.1
python-dotenv==1.0.1
Requests==2.32.3
supabase==2.6.0
tiktoken==0.7.0
lxml[html_clean]==4.9.3